Demo accomodations
Modified demo docker images to utilize unofficial build of Spark 2.4.5 for Hadoop 2.8.5 in
order to utilize S3A filesystem.  Attempting this with 2.7.3 fails due to a lack of support
for specifying byte sizes with a unit character (e.g. 512m or 2g) that was added post-2.7.3,
but is used for defaults acquired from the 2.8.5 Hadoop install.

Modified demo docker-compose to add Schema Registry, Connect, and a target Postgres database
in order to use the infrastructure to demonstrate using Hudi with Debezium CDC.

Modified demo docker images to download the latest HUDI relase jar rather than mounting the
workspace directory in order to yield a compose document that is compatible with extraction
for Kubernetes by kompose.

Modified demo docker images to expose a smaller number of ports for fewer unnecessarily
exposed ports

Exchanged Bitnami Zookeeper and Kafka images for Debezium's and Confluent's in order to lay
groundwork for running the demo with a multi-broker container base.  ** Note that this is
not yet leveraged--still only one Kafak broker is launched.  *** Making this change preemptively
did, however, make it easier to port my earlier work with Connect and Schema registry in, and
that very much is used at present.