pom.xml dependency updates
Potential reversion here!

Not sure if it is necessary for HUDI to build with Spark 2.4.5 and Hadoop 2.8.5
dependencies for the demo changes explored to work, and I highly suspect that it
is not needed at all.  Reason being that the change from depending on the current
workspace view to using the latest released HUDI jar seem to work fine and do not
reflect these changes.